finished the admin nav bar
