+++ src/pages/AdminHome.jsx
@@ -4,6 +4,7 @@
 import React, { useState} from "react"
 import { useNavigate } from "react-router-dom";
 import NavBar from "../components/NavBar";
+import AdminNavBar from "../components/AdminNavBar.jsx";
 import OrangeIcon from "../assets/ExploreByAge.jpg";
 import BlueIcon from "../assets/ExploreByGenre.jpg";
 import YellowIcon from "../assets/NewArrivalsIcon.jpg";
@@ -23,11 +24,6 @@ export default function AdminHome({}) {
     const navigate = useNavigate(); 
     
     //handle routes
-    const handleSearch = () => {
-      console.log("HOME.jsx searchInput: " , searchInput);
-      navigate("/search", { state: { initSearchInput: searchInput }});
-    };
-  
     const handleSuggestBookNav = () => {
       navigate("/suggest");
     };
@@ -39,12 +35,8 @@ export default function AdminHome({}) {
     const handleExploreByGenre = () => {
       setShowPulloutBarGenre(!showPulloutBarGenre)
     };
-  
-    //event 
-    const handleKeyDown = (event) => {
-      if (event.key === 'Enter') {
-        handleSearch()
-      }
+    const handleTestClick = () => {
+      console.log('Button pressed');
     };
 
   return (
@@ -52,20 +44,14 @@ export default function AdminHome({}) {
 
   
 <div className={`h-screen w-screen pb-5 start-bg flex flex-col items-center ${isBlurred ? 'blur-sm' : ''}`}>
-    <NavBar useDarkTheme={false} showTitle={false} bgColor={"#110057"} textColor={"#FFFFFF"} />
-    <h1 className="absolute left-40 top-1/6 mt-16 text-5xl text-white">Bibliotrace 3.0</h1>
-    {/* Search Bar */}
-    <div className="absolute left-14 top-28 h-16 my-6 flex w-1/2 justify-center"> 
-      <input className="m-2 px-3 w-10/12 border-2 border-[#a49bc6] rounded-2xl placeholder-[#a49bc6] placeholder:font-bold" type="text" placeholder="Search" value={searchInput} onInput={e => setSearchInput(e.target.value)} onKeyDown={handleKeyDown}></input>
-      <button className="m-2 border-[#110057] border-2 bg-white rounded-2xl font-bold text-[#a49bc6]" onClick= {handleSearch}>Go!</button>
-    </div>
+  <AdminNavBar/>
     
-    <div className="absolute left-0 top-48 flex flex-row">
+    <div className="absolute left-0 top-96  flex flex-row">
       <CustomButton
           imageSrc={PurpleIcon}
           text="CHECK OUT"
           textColor="#FFFFFF"
-          onClick={handleSearch}
+          onClick={handleTestClick}
           borderColor="#4b00e3" 
           bgColor="#110057"
           layout='row'
@@ -77,7 +63,7 @@ export default function AdminHome({}) {
           imageSrc={OrangeIcon}
           text="CHECK IN"
           textColor="#FFFFFF"
-          onClick={handleExploreByAge}
+          onClick={handleTestClick}
           borderColor="#fa8804" 
           bgColor="#110057"
           layout='row'
@@ -87,7 +73,7 @@ export default function AdminHome({}) {
           imageSrc={PinkIcon}
           text={"POPULAR"}
           textColor="#FFFFFF"
-          onClick={handleSearch}
+          onClick={handleExploreByAge}
           borderColor="#ff50df" 
           bgColor="#110057"
           layout='row'
@@ -120,7 +106,7 @@ export default function AdminHome({}) {
           imageSrc={YellowIcon}
           text={"NEW ARRIVALS"}
           textColor="#FFFFFF"
-          onClick={handleSearch}
+          onClick={handleTestClick}
           borderColor="#FFD700" 
           bgColor="#110057"
           layout='row'
@@ -142,21 +128,21 @@ export default function AdminHome({}) {
         <CustomTextBoxButton
             text= "Manage Inventory" 
             textColor = "#110057" 
-            onClick={handleSearch} 
+            onClick={handleTestClick} 
             borderColor = "#110057" 
             bgColor= "#FFFFFF" 
         />
         <CustomTextBoxButton
             text= "Manage Inventory" 
             textColor = "#110057" 
-            onClick={handleSearch} 
+            onClick={handleTestClick} 
             borderColor = "#110057" 
             bgColor= "#FFFFFF" 
         />
         <CustomTextBoxButton
             text= "Manage Inventory" 
             textColor = "#110057" 
-            onClick={handleSearch} 
+            onClick={handleTestClick} 
             borderColor = "#110057" 
             bgColor= "#FFFFFF" 
         />
